Extract GitHub runtime account scope helpers
